% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{bus_query}
\alias{bus_query}
\title{Function to query service, ewt, wtp, otp endpoints}
\usage{
bus_query(endpoint = "speed", year_list = list(), month_list = list(),
  route_list = list(), stop_list = list(), weekend = NA,
  period_list = list(), sbs = NA, group_list = list(),
  order_list = list(), limit = NA, token = NA)
}
\arguments{
\item{endpoint}{String. Endpoint to query one of: service, ewt, wtp, otp}

\item{year_list}{Integer vector of years to query (2014, 2015, or 2016)}

\item{month_list}{Character vector of months to query in YYYYMMDD or YYYY-MM-DD format}

\item{route_list}{Character vector of route ids to query}

\item{stop_list}{Character vector of stop ids to query}

\item{weekend}{T/F flag for analyzing weekends. True is just weekends, false is no weekends. Not passing is all days}

\item{period_list}{Integer vector of periods to query. One or more of:
1: 7am–10am
2: 10am–4pm
3: 4pm–7pm
4: 7pm–11pm
5: 11pm–7am}

\item{sbs}{T/F flag for whether to include SBS routes in computations}

\item{group_list}{Character vector of fields to aggregate by. One or more of:
year
month
route_id
stop_id
weekend
period}

\item{order_list}{Character vector of fields to sort by. One of: year, month, route_id, stop_id, weekend, period, or
one of the metrics returne from endpoint. See API documentation. Add a hyphen to field to sort in descending order
(e.g. -year)}

\item{limit}{Integer. Number of results to return}

\item{token}{String. API documentation says this is required but doesn't appear to be so. See API page to obtain one.}
}
\value{
a dataframe of results
}
\description{
Function to query service, ewt, wtp, otp endpoints
}
\examples{
bus_query(endpoint = "speed", year_list = c(2015), weekend = FALSE, period_list = c(1,2,3),
group_list = ("route_id"), order_list = c("speed"), limit = 10)

}

